pyproject.toml
[project]
name = "python-orm-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "ipykernel>=6.29.5",
    "pgvector>=0.4.1",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.0",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.3",
]
